+++ src/pages/auth/signIn/index.tsx
@@ -5,58 +5,49 @@ import { useRouter } from 'next/router';
 import { useEffect, useState } from 'react';
 import { useAuthState } from 'react-firebase-hooks/auth';
 
-import { auth, registerWithEmailAndPassword, signInWithGoogle } from '@/app/components/FireBase';
+import { auth, logInWithEmailAndPassword, signInWithGoogle } from '@/app/components/FireBase';
 import { Sidebar } from '@/widgets/layouts/side-bar';
 
-const SignUp = () => {
+const SignIN = () => {
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
-  const [name, setName] = useState('');
   const [user, loading, error] = useAuthState(auth);
   const router = useRouter();
-
-  const register = () => {
-    if (!name) alert('Please enter name');
-    registerWithEmailAndPassword(name, email, password);
-  };
   useEffect(() => {
-    if (loading) return;
-    if (user) router.push('/');
+    if (loading) {
+      // maybe trigger a loading screen
+      return;
+    }
+    if (user) router.push('/main');
   }, [user, loading, router]);
   return (
     <>
       <Sidebar />
-      <div className="register">
-        <div className="register__container">
+      <div className="login">
+        <div className="login__container">
           <TextField
             type="text"
-            className="register__textBox"
-            value={name}
-            onChange={(e) => setName(e.target.value)}
-            placeholder="Full Name"
-          />
-          <TextField
-            type="text"
-            className="register__textBox"
             value={email}
             onChange={(e) => setEmail(e.target.value)}
-            placeholder="E-mail Address"
+            placeholder="E-mail"
           />
           <TextField
             type="password"
-            className="register__textBox"
             value={password}
             onChange={(e) => setPassword(e.target.value)}
             placeholder="Password"
           />
-          <Button variant="contained" onClick={register}>
-            Register
+          <Button variant="contained" onClick={() => logInWithEmailAndPassword(email, password)}>
+            SignIn
           </Button>
           <Button variant="contained" onClick={signInWithGoogle}>
-            Register with Google
+            SignIn with Google
           </Button>
           <div>
-            Already have an account? <Link href="/signIn">Login</Link> now.
+            <Link href="/reset">Forgot Password</Link>
+          </div>
+          <div>
+            Do not have an account? <Link href="/auth/signUp">Register</Link> now.
           </div>
         </div>
       </div>
@@ -64,4 +55,4 @@ const SignUp = () => {
   );
 };
 
-export default SignUp;
+export default SignIN;
